fix: move system status into features
