+++ src/ui/layouts/MainLayout.tsx
@@ -2,7 +2,7 @@ import { Outlet } from "react-router-dom";
 import { Footer } from "../components/common/footer/Footer";
 import { Header } from "./header/Header";
 import { MenuProvider } from "../providers/MenuProvider";
-import { ServerStatus } from "../components/server-status/ServerStatus";
+import { ServerStatus } from "../features/system/components/server-status/ServerStatus";
 
 import "./MainLayout.css";
