+++ src/ui/pages/settings/ServerSettings.tsx
@@ -2,7 +2,7 @@ import { BaseSettings } from "./BaseSettings";
 import { useUserSettings } from "../../providers/UserSettingsProvider";
 import { Stack, Paper, Group, Badge, Divider, Alert } from "@mantine/core";
 import { IconInfoCircle } from "@tabler/icons-react";
-import { ServerStatus } from "../../components/server-status/ServerStatus";
+import { ServerStatus } from "../../features/system/components/server-status/ServerStatus";
 
 export function ServerSettings() {
     // Use settings.
